bot hosting/requirements.txt
python-telegram-bot==13.15
flask==2.3.2
gunicorn==20.1.0
requests==2.31.0
zipfile36==0.1.4